+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### Fixed
 
-- TableView.onEditStart() and TableColumn.cancel() functions which can be used to intercept editing events
-
 ### Changes
 
 ### Additions
+
+- TableView.onEditStart() and TableColumn.cancel() functions which can be used to intercept editing events
+- resources.media() to load a Media instance from resources
+- Media.play() shortcut which creates a MediaPlayer and plays the Media
 
 ## [1.7.17]